Refoctore code
