+++ package.json
@@ -11,11 +11,13 @@
   },
   "dependencies": {
     "@radix-ui/react-dropdown-menu": "^2.1.15",
+    "@radix-ui/react-label": "^2.1.7",
     "@radix-ui/react-select": "^2.2.5",
     "@radix-ui/react-slot": "^1.2.3",
     "apexcharts": "^4.7.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
+    "jotai": "^2.12.5",
     "lucide-react": "^0.514.0",
     "react": "^19.1.0",
     "react-apexcharts": "^1.7.0",
