+++ src/charts/dataCharts.jsx
@@ -1,0 +1,4 @@
+import { atom } from "jotai";
+
+export const rawDataAtom = atom(null);
+export const chartDataAtom = atom(null);
