--- src/components/Context.jsx
@@ -1,5 +1,0 @@
-import { createContext, useContext } from "react";
-
-export const FormContext = createContext(null);
-
-export const useFormContext = () => useContext(FormContext);
